+++ CHANGES.rst
@@ -8,6 +8,7 @@ Unreleased
 -   Drop support for Python 2.7 and 3.5.
 -   Bump MarkupSafe dependency to >=1.1.
 -   Allow custom autoescape functions :issue:`1377`
+-   Removed hardcoded Markup and escape calls
 -   Bump Babel optional dependency to >=2.1.
 -   Remove code that was marked deprecated.
 -   Use :pep:`451` API to load templates with